+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## Release 4.7
 
 * #488 Bump to the latest version of openapi-generator (v5.2.0)
+* #490 "createOrUpdate" and "idempotentCreateOrUpdate" must return 201 when a single object is created
 
 ## Release 4.6